Replaces stationResampler with electricFieldResampler
